ATO-NONE: Move account intervention service to shared location
