+++ orchestration-shared/src/main/java/uk/gov/di/orchestration/shared/exceptions/AccountInterventionException.java
@@ -1,4 +1,4 @@
-package uk.gov.di.authentication.oidc.exceptions;
+package uk.gov.di.orchestration.shared.exceptions;
 
 public class AccountInterventionException extends RuntimeException {
     public AccountInterventionException(String message, Exception cause) {
